Ensure the simultaneous drawing of two squares
To achieve the simultaneous rendering of two squares, distinct
coordinates were assigned to each square. This enables the depiction of
both squares at distinct positions, preventing any conflict in their
attributes.

Previously, there existed a sole square. It might have appeared as if
there were two squares due to alterations in the coordinates and color
of the same square, albeit the square was essentially being repositioned.

Presently, two separate squares are generated, each possessing their
unique coordinates and colors. Furthermore, the drawSquares function has
been modified to render both squares concurrently.

All movement functions and values now reference the values of square1,
for example, what was previously just "x" is now "x1", "speed" became
"speed1", etc.

This solution is far from optimal, there is a bunch of opportunity to
improve the code. However, we might not be at that stage yet.

A recommendation moving forward is to try and structure/organise the
code. It will help you when debugging as well as understanding what's
going on.